refactor: 패키지 이동
- DispatcherServletInitializer 패키지를 app 으로 이동
